requirements.txt
Django==4.1.5
django-autoslug==1.9.8
psycopg2-binary==2.9.5
